pyproject.toml
[tool.poetry]
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.20.0.post0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
alembic = "^1.15.2"
asyncpg = "^0.30.0"
pydantic-settings = "^2.9.1"


[tool.poetry.group.lint.dependencies]
ruff = "^0.11.10"
black = "^25.1.0"
pre-commit = "^4.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"


[tool.black]
line-length = 110


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".*_factory.py",
]

# Same as Black.
line-length = 110

# Assume Python 3.11
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.

select = [
    "F", # Pyflakes
    "E", # Error
    "E4", # Error
    "E9", # Error
    "I", # isort
    "BLE", # flake8-blind-except
    "YTT", # flake8-2020
    "ASYNC", # flake8-async
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
